Renaming parsers to clients and moving files. Adding email validator to pydantic validation
