+++ tests/clients/test_crossref.py
@@ -3,7 +3,7 @@ import re
 import pytest
 
 import credit_engine.constants as CE
-from credit_engine.parsers.crossref import get_endpoint, retrieve_doi
+from credit_engine.clients.crossref import get_endpoint, retrieve_doi
 from tests.common import check_stdout_for_errs
 from tests.conftest import (
     GET_ENDPOINT_FAIL_DATA,
@@ -66,9 +66,8 @@ GET_ENDPOINT_DATA = [
         {
             "input": [SAMPLE_DOI, "UNIXSD", ""],
             "error": re.escape(
-                "1 validation error for GetEndpoint\nemail_address\n  ensure "
-                "this value has at least 1 characters "
-                "(type=value_error.any_str.min_length; limit_value=1)"
+                "1 validation error for GetEndpoint\nemail_address\n  value is not a valid email address "
+                "(type=value_error.email)"
             ),
         },
         id="fmt_unixsd_email_len_0",
@@ -77,12 +76,21 @@ GET_ENDPOINT_DATA = [
         {
             "input": [SAMPLE_DOI, "UNIXSD", SPACE_STR],
             "error": re.escape(
-                "1 validation error for GetEndpoint\nemail_address\n  ensure "
-                "this value has at least 1 characters "
-                "(type=value_error.any_str.min_length; limit_value=1)"
+                "1 validation error for GetEndpoint\nemail_address\n  value is not a valid email address "
+                "(type=value_error.email)"
             ),
         },
         id="fmt_unixsd_email_whitespace",
+    ),
+    pytest.param(
+        {
+            "input": [SAMPLE_DOI, "UNIXSD", "fake email address"],
+            "error": re.escape(
+                "1 validation error for GetEndpoint\nemail_address\n  value is not a valid email address "
+                "(type=value_error.email)"
+            ),
+        },
+        id="fmt_unixsd_email_fake_email",
     ),
     pytest.param(
         {
