+++ tests/clients/test_osti.py
@@ -1,7 +1,7 @@
 import pytest
 
 import credit_engine.constants as CE
-from credit_engine.parsers.osti import get_endpoint, retrieve_doi
+from credit_engine.clients.osti import get_endpoint, retrieve_doi
 from tests.common import check_stdout_for_errs
 from tests.conftest import (
     GET_ENDPOINT_FAIL_DATA,
